kyc/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
insightface==0.7.3
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
PaddleOCR==2.7.0.3
Pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.1
pydantic==1.10.15
